+++ windows/v60_viiper_app/MainWindow.xaml.cs
@@ -62,7 +62,7 @@ public partial class MainWindow : Window
         trayIcon = new Forms.NotifyIcon
         {
             Icon = SystemIcons.Application,
-            Text = "PRO2 控制板 V6.2.0 新和联胜",
+            Text = "PRO2 控制板 V6.2.4 新和联胜",
             ContextMenuStrip = trayMenu,
             Visible = true
         };